+++ object/expectation/json_marshaler.go
@@ -1,0 +1,20 @@
+package expectation
+
+import (
+	"encoding/json"
+)
+
+func (e *expectation) MarshalJSON() ([]byte, error) {
+	type ExpectationClone expectation
+
+	b, err := json.Marshal(&struct {
+		*ExpectationClone
+	}{
+		ExpectationClone: (*ExpectationClone)(e),
+	})
+	if err != nil {
+		return nil, maskAny(err)
+	}
+
+	return b, nil
+}
