+++ object/expectation/json_unmarshaler.go
@@ -1,0 +1,21 @@
+package expectation
+
+import (
+	"encoding/json"
+)
+
+func (e *expectation) UnmarshalJSON(b []byte) error {
+	type ExpectationClone expectation
+
+	aux := &struct {
+		*ExpectationClone
+	}{
+		ExpectationClone: (*ExpectationClone)(e),
+	}
+	err := json.Unmarshal(b, &aux)
+	if err != nil {
+		return maskAny(err)
+	}
+
+	return nil
+}
