+++ service/permutation/permutation_test.go
